requirements.txt
asgiref==3.5.2
Django==4.1.2
django-cors-headers==3.13.0
django-rest-framework==0.1.0
djangorestframework==3.14.0
gunicorn==20.1.0
Pillow==9.3.0
psycopg2-binary==2.9.3
pytz==2022.4
sqlparse==0.4.3

